poolbet-bot/requirements.txt
python-telegram-bot[job-queue]>=20.7
asyncpg>=0.29.0
web3>=6.15.0
eth-account>=0.10.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
Pillow>=10.2.0
redis>=5.0.1
hiredis>=2.3.2
requests>=2.31.0
